finish review gis public functions
