+++ hydromt/gis/vector_utils.py
@@ -17,12 +17,12 @@ logger = logging.getLogger(__name__)
 
 __all__ = [
     "_nearest",
-    "_nearest_merge",
     "_filter_gdf",
+    "nearest_merge",
 ]
 
 
-def _nearest_merge(
+def nearest_merge(
     gdf1: gpd.GeoDataFrame,
     gdf2: gpd.GeoDataFrame,
     *,
@@ -50,8 +50,6 @@ def _nearest_merge(
         i.e. NaN values for existing columns or missing columns.
     inplace : bool,
         If True, apply the merge to gdf1, otherwise return a new object.
-    logger:
-        The logger to use.
 
     Returns
     -------
@@ -63,15 +61,25 @@ def _nearest_merge(
     if not inplace:
         gdf1 = gdf1.copy()
     valid = dst < max_dist if max_dist is not None else np.ones_like(idx_nn, dtype=bool)
+    # gdf2 column selection
+    columns = gdf2.columns if columns is None else columns
+    for c in columns:
+        if c not in gdf2.columns:
+            logger.warning(f"Column {c} not in gdf2, skipping.")
+            columns.remove(c)
+
     gdf1["distance_right"] = dst
     gdf1["index_right"] = -1
     gdf1.loc[valid, "index_right"] = idx_nn[valid]
 
     if not overwrite:
-        new_cols = [c for c in gdf2.columns if c not in gdf1.columns]
+        new_cols = [c for c in columns if c not in gdf1.columns]
         gdf1.loc[:, new_cols] = np.nan
         return gdf1.combine_first(gdf2)
     else:
+        # Keep only columns selection in gdf2 before join
+        gdf2 = gdf2[columns]
+        # Keep only left only columns in gdf1 before join
         left_only_cols = [c for c in gdf1.columns if c not in gdf2.columns]
         return gdf1[:, left_only_cols].join(gdf2, join="left")
